requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
azure-ai-formrecognizer==3.3.2
azure-core==1.29.6
openai==1.12.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
ultralytics==8.1.9
torch==2.2.0
torchvision==0.17.0
scikit-image==0.22.0
scikit-learn==1.4.0
Pillow==10.2.0
python-multipart==0.0.6
loguru==0.7.2
httpx==0.26.0
